% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment_table.R
\name{adjustment_table}
\alias{adjustment_table}
\title{Extract adjustment table}
\usage{
adjustment_table(DAG)
}
\arguments{
\item{DAG}{tidy_dagitty object. Generated by \code{draw_dag}.}
}
\value{
A tibble containing following columns:
}
\description{
Using a DAG information, extracts adjustment tables for
pre-specified (hard-coded) outcome/predictor pairs.
}
