% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_means.R
\name{compare_means}
\alias{compare_means}
\title{Compare marginal means}
\usage{
compare_means(means, specs)
}
\arguments{
\item{means}{List. Expected marginal means such as
those computed by \code{compute_means}.}

\item{specs}{Tibble or data.frame. Contains research
questions operationalisation.}
}
\value{
A tibble with pairwise comparisons.
}
\description{
Given expected marginal means (e.g., calculated via
\code{compute_means}) and contrast specifications,
}
