% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_means.R
\name{compute_means}
\alias{compute_means}
\title{Compute marginal means}
\usage{
compute_means(fits, specs)
}
\arguments{
\item{fits}{List. Model fits such as those generated
by \code{fit_models}.}

\item{specs}{Tibble or data.frame. Contains research
questions operationalisation.}
}
\value{
Named list containing \code{emmeans} objects with
marginal means of interest per model.
}
\description{
Given model fits and contrast specifications,
computes expected marginal means of grouping
of interest.
}
