% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_models.R
\name{diagnose_models}
\alias{diagnose_models}
\title{Plot model diagnostics}
\usage{
diagnose_models(model_list)
}
\arguments{
\item{model_list}{List. Regression models to check.}
}
\value{
A list with diagnostic plots of the same nested
structure as \code{model_list}.
}
\description{

}
