% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_dag.R
\name{draw_dag}
\alias{draw_dag}
\title{Draw a directed acyclic graphs}
\usage{
draw_dag(plot = TRUE)
}
\arguments{
\item{plot}{Logical. Should ggplot be returned (\code{TRUE}, default) or
tidy_daggity object (\code{FALSE}).}
}
\value{
Directed acyclic graph representing causal assumptions
for the analysis.
}
\description{
Takes in sets of covariates derived by \code{adjustment_table}
and changes them to equations that can be used as \code{formula}.
}
