% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_means.R
\name{extract_means}
\alias{extract_means}
\title{Extract marginal means}
\usage{
extract_means(means, specs, digits = 2)
}
\arguments{
\item{means}{List. Expected marginal means such as
those computed by \code{compute_means}.}

\item{specs}{Tibble or data.frame. Contains research
questions operationalisation.}

\item{digits}{Integer. Decimal places.}
}
\value{
A tibble with formatted marginal means.
}
\description{
Extracts and formats marginal means computed
by \code{compute_means}.
}
