% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norms.R
\name{extract_thresholds}
\alias{extract_thresholds}
\title{Find memory thresholds for superaging}
\usage{
extract_thresholds(
  pvlt_data_file,
  pvlt_id_file,
  nanok_file,
  ravlt_file,
  output = "thresholds"
)
}
\arguments{
\item{pvlt_data_file}{Character. Path do PVLT normative
data.}

\item{pvlt_id_file}{Character. Path do PVLT included IDs.}

\item{nanok_file}{Character. Path do NANOK normative
data.}

\item{ravlt_file}{Character. Path do RAVLT normative
data (subset of NANOK).}

\item{output}{Character. Return thresholds for SuperAging
(\code{"thresholds"}) or normative values (\code{"norms"}).}
}
\value{
RAVLT and PVLT thresholds for SuperAging or
normative tables of RAVLT and PVLT.
}
\description{
Using normative values of RAVLT and PVLT raw normative
data.
}
