% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{Fit regression models}
\usage{
fit_models(data, specs, log_1 = c("GDS15", "GAI"), contr = TRUE)
}
\arguments{
\item{data}{Tibble or data.frame.}

\item{specs}{Tibble or data.frame.}

\item{log_1}{Character. Which variables \code{y} should be
analysed as their \code{log(y+1)} versions? Defaults to
\code{c("GDS15","GAI")}.}

\item{contr}{Logical. Set sum contrasts to factors to
avoid multicollinearity? Defaults to \code{TRUE}.}
}
\value{
A named nested list with regression models.
}
\description{
Using table of adjustments derived via
\code{adjustment_table}, this function
prepares model specifications for later
fitting.
}
