% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import data for analysis}
\usage{
import_data(file, sheet, norms, thresholds, thres_type = "mean")
}
\arguments{
\item{file}{Character. Path to raw data XLSX file.}

\item{sheet}{Character or numeric. Denotes sheet where the
data resides inside of \code{file}.}

\item{norms}{Tibble or data.frame. Normative values for scoring
memory tasks. Computed by \code{extract_thresholds}.}

\item{thresholds}{Tibble or data.frame. Thresholds based on normative
values of  memory tasks. Computed by \code{extract_thresholds}.}

\item{thres_type}{Character.}
}
\value{
A tibble with following columns:
}
\description{
Reads, pre-processes, and cleans data for analysis.
}
