% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_specs.R
\name{model_specs}
\alias{model_specs}
\title{Set models' specifications}
\usage{
model_specs(table, faq = "count")
}
\arguments{
\item{table}{Tibble. Generated by \code{adjustment_table}.}

\item{faq}{Character. Should FAQ be treated as binomial
("count", default) or continuous ("continuous")?}
}
\value{
A tibble with following columns:
}
\description{
Using tabale of adjustments derived via
\code{adjustment_table}, this function
prepares model specifications for later
fitting.
}
