% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{Plot results}
\usage{
plot_results(data, stats, specs, type, save = TRUE)
}
\arguments{
\item{data}{Tibble or data.frame. Raw data as generated
by \code{import_data}.}

\item{stats}{Tibble or data.frame. Statistic comparisons
as generated by \code{stat_test}.}

\item{specs}{Tibble or data.frame. Contains research
questions operationalisation.}

\item{type}{Character.}

\item{save}{Logical. Save the plot into \code{figures} folder?
Defaults to \code{TRUE} which also creates \verb{figures/} if it
does not exist yet.}
}
\value{
A tibble with formatted marginal means and their
comparisons.
}
\description{
Plots raw data and their summaries against regression
results, including verbal description of the results.
}
