% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norms.R
\name{ravlt_ss}
\alias{ravlt_ss}
\title{Compute scaled scores}
\usage{
ravlt_ss(raw)
}
\arguments{
\item{raw}{Numeric. Raw RAVLT delayed recall score.}
}
\value{
Numeric vector with scaled scores.
}
\description{
Using Table 7 from the normative study by Frydrychova et al.
(\href{https://psycnet.apa.org/record/2018-63633-003}{2018}), this
function maps a single raw RAVLT delayed recall score to a
scaled score (M = 10, SD = 3).
}
