% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_brackets}
\alias{remove_brackets}
\title{Remove set brackets}
\usage{
remove_brackets(x)
}
\arguments{
\item{x}{Character.}
}
\value{
A character with curly brackets removed.
}
\description{
Remove set brackets
}
