% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rprint}
\alias{rprint}
\title{Print rounded number}
\usage{
rprint(x, d = 2)
}
\arguments{
\item{x}{Numeric. Number to be rounded.}

\item{d}{Numeric. Decimal places (defaults to 2).}
}
\value{
A character with \code{d} decimals.
}
\description{
Print rounded number
}
