% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set2equation.R
\name{set2equation}
\alias{set2equation}
\title{Change sets to equations}
\usage{
set2equation(x, bracket = FALSE)
}
\arguments{
\item{x}{Character. Contains sets to be transformed.}

\item{bracket}{Logical. Bracket covariates into interactions?
Defaults to \code{FALSE}.}
}
\value{
Character vector with components that can be changed
to formulas.
}
\description{
Takes in sets of covariates derived by \code{adjustment_table}
and changes them to equations that can be used as \code{formula}.
}
