% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_test.R
\name{stat_test}
\alias{stat_test}
\title{Table results}
\usage{
stat_test(fits, specs, sets)
}
\arguments{
\item{fits}{List. Model fits such as those generated
by \code{fit_models}.}

\item{specs}{Tibble or data.frame. Contains research
questions operationalisation.}

\item{sets}{Tibble. Contains variable sets mapping to
table labels. E.g., generated by \code{adjustment_table}.}
}
\value{
A tibble with formatted marginal means and their
comparisons.
}
\description{
Computes marginal means and their comparisons of
a set of regression models and tables the results.
}
