% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zerolead}
\alias{zerolead}
\title{Drop leading zero}
\usage{
zerolead(x, d = 3)
}
\arguments{
\item{x}{Numeric. Number to be rounded.}

\item{d}{Numeric. Decimal places (defaults to 3).}
}
\value{
A character with \code{d} decimals and no leading
zero. If \code{x < 0.001}, the \code{"< .001"} is printed.
}
\description{
Drop leading zero
}
